+++ src/game_state.h
@@ -190,7 +190,7 @@ class GameState
 
 		// HUD
 		void addHUDMessage(unsigned int slot, string text, string text2 = "");
-		HUDLabel* addHUDLabel(unsigned int slot, float x, float y, string data, HUDLabel* l = NULL);
+		HUDLabel* addHUDLabel(unsigned int slot, int x, int y, string data, HUDLabel* l = NULL);
 		bool mousePick(unsigned int x, unsigned int y, btVector3& hitLocation, Entity** hitEntity);
 
 		// Debugging
